Add function to create "diffs" dataset
- New function `diff_datasets()` creates diff datasets for every .csv dataset in the `data/` directory as described in ccodwg/CovidTimelineCanada#20

Only potentially unexpected behaviour occurs when a dataset (or part of a dataset) is updated but does not increment the date (i.e., only historical information are updated). This occurs sometimes with the vaccine datasets from PHAC, as the numbers are only updated every few weeks but old numbers are sometimes updated on off-weeks. For now, the behaviour counts this as a diff with 0 days between dates, which should be interpreted as an update of historical data. The alternative would be to replace the current data value only and simply re-do the diff calculation, but it is quite possible that the previous week's value has changed as well, rendering the calculation unreliable. I suppose a potential solution would be to also update the previous value using the new dataset, but this might be a little much.